Don't initialise verb list with language
Initialise and store the verb list only with the inspector window, as it
is only needed there. This might degrade performance a little bit, but
it alleviates huge headaches with linking order on different platforms,
as now all the different verbs don't necessarily need to be defined
before the languages are initialised.

Doing it like this might even improve memory usage a little bit at times
where the inspector window is not open, which is most of the time.